Create Guessing Challenging Game
Guessing Challenging Game is an old game in which at least there are two players. 
One will select a number from a specified range and the other should guess . As the second player gets closer to number or gets further , the first player will notify him/her based on agreed rules. 
Now the same game is written in Python , so the game will choose a number and you have to guess the number. 
 Note: that there is no limitation of guesses have been choose in code but if you like you can add it yourself.